encapsulate API logic into wrapper component
