+++ frontend/src/components/form/ApiTextField.vue
@@ -4,72 +4,32 @@ You can two-way bind to the value using v-model.
 -->
 
 <template>
-  <span>
-    <span v-if="!editing">{{ value }}</span>
-    <span v-if="editing">
-      <v-form
-        inline
-        class="mb-2">
-
-        <v-text-field
-          id="api-text-field"
-          v-model="localValue"
-          :label="label"
-          name="api-text-field"
-          class="mr-2 ml-2"
-          :error-messages="errorMessage"
-          :state="required && $v.localValue.$dirty ? !$v.localValue.$error : null"
-          v-bind="$attrs"
-          required
-          @input="onInput"
-          @blur="$v.localValue.$touch()">
-
-          <template slot="append-outer">
-            <v-btn
-              v-if="!autoSave"
-              small
-              color="warning"
-              class="mb-1"
-              @click="reset">
-
-              Reset
-            </v-btn>
-
-            <v-btn
-              color="primary"
-              small
-              :disabled="isSaving || (required && $v.localValue.$invalid)"
-              class="mr-2 ml-2 mb-1"
-              @click="save">
-
-              <v-progress-circular
-                v-if="isSaving"
-                indeterminate
-                color="primary"
-                size="20"
-                class="mr-2" />
-
-              <v-icon
-                v-if="showSuccessIcon"
-                left>
-                mdi-check</v-icon>
-
-              Save
-            </v-btn>
-          </template>
-        </v-text-field>
-
-      </v-form>
-    </span>
-  </span>
+  <api-wrapper
+    v-bind="$props">
+    <template slot-scope="wrapper">
+      <v-text-field
+        :value="wrapper.localValue"
+        v-bind="$attrs"
+        :label="label"
+        :error-messages="wrapper.errorMessage"
+        @input="wrapper.on.input"
+        @blur="wrapper.on.touch">
+        <status-icon slot="append" :status="wrapper.status" />
+      </v-text-field>
+    </template>
+  </api-wrapper>
 </template>
 
 <script>
-import { apiMixin } from '@/mixins/apiMixin'
+import { apiPropsMixin } from '@/mixins/apiPropsMixin'
+import ApiWrapper from './ApiWrapper'
+import StatusIcon from './StatusIcon'
 
 export default {
   name: 'ApiTextField',
-  mixins: [apiMixin],
+  components: { ApiWrapper, StatusIcon },
+  mixins: [apiPropsMixin],
+
   data () {
     return {}
   }
@@ -77,7 +37,4 @@ export default {
 </script>
 
 <style scoped>
-  .dirty{
-    border:1px red solid;
-  }
 </style>
