+++ frontend/src/mixins/apiPropsMixin.js
@@ -1,0 +1,28 @@
+export const apiPropsMixin = {
+  inheritAttrs: false,
+
+  props: {
+    value: { type: String, required: true },
+
+    /* field name and URI for saving back to API */
+    fieldname: { type: String, required: true },
+    uri: { type: String, required: true },
+
+    /* display label */
+    label: { type: String, default: '', required: false },
+
+    /* overrideDirty=true will reset the input if 'value' changes, even if the input is dirty. Use with caution. */
+    overrideDirty: { type: Boolean, default: false, required: false },
+
+    /* enable/disable edit mode */
+    editing: { type: Boolean, default: true, required: false },
+
+    /* enable/disable auto save */
+    autoSave: { type: Boolean, default: true, required: false },
+    autoSaveDelay: { type: Number, default: 800, required: false },
+
+    /* Validation criteria */
+    required: { type: Boolean, default: false, required: false }
+  }
+
+}
